Replaced projects path with profile path
